cli/pyproject.toml
[tool.poetry.dependencies]
python = "^3.8"
typer = "^0.16.0"
mpy-cross = "^1.25.0"
pyyaml = "^6.0.0"
requests = "^2.32.4"
pyserial = "^3.5"
mpremote = "^1.25.0"
beautifulsoup4 = "^4.13.4"
psutil = "^7.0.0"
tqdm = "^4.67.1"